Simpllify files fo ConsentScreen, WelcomeScreen 1 & 2 and the CountryIpModal to follow our repo style
