+++ src/CovidApp.tsx
@@ -42,11 +42,11 @@ import PreviousExposureScreen from '@covid/features/patient/PreviousExposure';
 import YourHealthScreen from '@covid/features/patient/YourHealthScreen';
 import YourStudyScreen from '@covid/features/patient/YourStudyScreen';
 import YourWorkScreen from '@covid/features/patient/YourWorkScreen';
-import { ConsentScreen } from '@covid/features/register/ConsentScreen';
+import ConsentScreen from '@covid/features/register/ConsentScreen/ConsentScreen';
 import { OptionalInfoScreen } from '@covid/features/register/OptionalInfoScreen';
 import { RegisterScreen } from '@covid/features/register/RegisterScreen';
-import { Welcome1Screen } from '@covid/features/register/Welcome1Screen';
-import { Welcome2Screen } from '@covid/features/register/Welcome2Screen';
+import Welcome1Screen from '@covid/features/register/Welcome1Screen';
+import Welcome2Screen from '@covid/features/register/Welcome2Screen';
 import { WelcomeRepeatScreen } from '@covid/features/register/WelcomeRepeatScreen';
 import { PrivacyPolicyUKScreen } from '@covid/features/register/gb/PrivacyPolicyUKScreen';
 import ValidationStudyConsentScreen from '@covid/features/register/gb/ValidationStudyConsentScreen';
