--- src/features/register/ConsentScreen/index.tsx
@@ -1,2 +1,0 @@
-import ConsentScreen from './ConsentScreen';
-export { ConsentScreen };
