--- src/features/register/ConsentScreen/styles.ts
@@ -1,16 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { colors } from '@theme';
-
-export default StyleSheet.create({
-  rootContainer: {
-    flex: 1,
-    justifyContent: 'space-between',
-    backgroundColor: colors.backgroundPrimary,
-    paddingHorizontal: 24,
-    paddingVertical: 24,
-  },
-  button: {
-    marginTop: 20,
-  },
-});
