--- src/features/register/CountryIpModal/index.tsx
@@ -1,2 +1,0 @@
-import CountryIpModal from './CountryIpModal';
-export default CountryIpModal;
