--- src/features/register/CountryIpModal/styles.ts
@@ -1,43 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { colors } from '@theme';
-
-const styles = StyleSheet.create({
-  centeredView: {
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginTop: 22,
-  },
-  modalView: {
-    margin: 24,
-    backgroundColor: colors.white,
-    borderRadius: 8,
-    padding: 24,
-    alignItems: 'center',
-    shadowColor: colors.black,
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-    elevation: 5,
-  },
-  titleText: {
-    marginVertical: 16,
-    fontSize: 20,
-    textAlign: 'center',
-  },
-  bodyText: {
-    textAlign: 'center',
-  },
-  labelStyle: {
-    fontSize: 15,
-    lineHeight: 30,
-    color: colors.primary,
-  },
-});
-
-export default styles;
