--- src/features/register/Welcome1Screen/index.tsx
@@ -1,5 +1,0 @@
-import { memo } from 'react';
-
-import Screen from './Welcome1Screen';
-
-export const Welcome1Screen = memo(Screen);
