+++ src/features/register/Welcome2Screen.tsx
@@ -1,22 +1,20 @@
 import { StackNavigationProp } from '@react-navigation/stack';
-import React, { FC, useState, useCallback } from 'react';
-import { Image, Linking, SafeAreaView, ScrollView, TouchableOpacity, View } from 'react-native';
+import React, { FC, useCallback, useState } from 'react';
+import { Image, Linking, SafeAreaView, ScrollView, StyleSheet, TouchableOpacity, View } from 'react-native';
 import { RouteProp } from '@react-navigation/native';
 
-import { usPartners, gbPartners, svPartners } from '@assets';
+import { gbPartners, svPartners, usPartners } from '@assets';
 import { BrandedButton, ClickableText, RegularBoldText, RegularText } from '@covid/components/Text';
-import UserService, { ICoreService, isGBCountry, isSECountry, isUSCountry } from '@covid/core/user/UserService';
+import { ICoreService, isGBCountry, isSECountry, isUSCountry } from '@covid/core/user/UserService';
 import { ScreenParamList } from '@covid/features/ScreenParamList';
 import i18n from '@covid/locale/i18n';
 import { useInjection } from '@covid/provider/services.hooks';
-import { IContentService } from '@covid/core/content/ContentService';
 import { Services } from '@covid/provider/services.types';
 import appCoordinator from '@covid/features/AppCoordinator';
+import { colors } from '@theme';
 
-import CountryIpModal from '../CountryIpModal';
-import { getLocaleFlagIcon } from '../helpers';
-
-import styles from './styles';
+import CountryIpModal from './CountryIpModal';
+import { getLocaleFlagIcon } from './helpers';
 
 const Slash = () => <RegularBoldText style={styles.slash}> / </RegularBoldText>;
 
@@ -137,5 +135,101 @@ const Welcome2Screen: FC<PropsType> = ({ navigation }) => {
     </SafeAreaView>
   );
 };
+
+const styles = StyleSheet.create({
+  safeView: {
+    flexGrow: 1,
+  },
+  scrollView: {
+    flexGrow: 1,
+    justifyContent: 'space-between',
+  },
+  rootContainer: {
+    flex: 1,
+    backgroundColor: colors.backgroundSecondary,
+  },
+  headerRow: {
+    paddingVertical: 8,
+    flexDirection: 'row',
+    justifyContent: 'flex-end',
+    alignItems: 'center',
+  },
+  covidContainer: {
+    paddingHorizontal: 24,
+    paddingBottom: 24,
+    paddingTop: 24,
+  },
+  partnerHeader: {
+    textAlign: 'center',
+    fontSize: 14,
+    lineHeight: 20,
+  },
+  divider: {
+    height: 1,
+    backgroundColor: colors.backgroundFour,
+    marginVertical: 5,
+  },
+  partnerList: {
+    marginTop: 0,
+    textAlign: 'center',
+    fontSize: 16,
+    lineHeight: 24,
+  },
+  login: {
+    color: colors.primary,
+    marginHorizontal: 16,
+  },
+  subheader: {
+    paddingVertical: 8,
+    color: colors.primary,
+    textAlign: 'center',
+    fontSize: 16,
+    lineHeight: 24,
+    marginTop: 16,
+  },
+  subheader2: {
+    paddingVertical: 8,
+    color: colors.secondary,
+    textAlign: 'center',
+    fontSize: 16,
+    fontWeight: '300',
+    lineHeight: 24,
+    marginTop: 8,
+  },
+  subtitle: {
+    color: colors.primary,
+    fontSize: 24,
+    lineHeight: 32,
+    paddingVertical: 8,
+    textAlign: 'center',
+    marginTop: 25,
+  },
+  slash: {
+    color: colors.lightBlueBrand,
+  },
+  partnersLogo: {
+    marginVertical: 16,
+    height: 160,
+    width: '100%',
+    resizeMode: 'contain',
+    alignSelf: 'center',
+  },
+  partnerContainer: {
+    marginVertical: 16,
+    paddingHorizontal: 30,
+    backgroundColor: colors.white,
+    borderRadius: 10,
+  },
+  buttonContainer: {
+    padding: 20,
+  },
+  flagIcon: {
+    height: 32,
+    width: 32,
+  },
+  nhsWebsite: {
+    textDecorationLine: 'underline',
+  },
+});
 
 export default Welcome2Screen;
