--- src/features/register/Welcome2Screen/index.tsx
@@ -1,5 +1,0 @@
-import { memo } from 'react';
-
-import Screen from './Welcome2Screen';
-
-export const Welcome2Screen = memo(Screen);
